+++ lib/oxidized/model/apcaos.rb
@@ -1,4 +1,4 @@
-class ApcOS < Oxidized::Model
+class ApcAos < Oxidized::Model
   using Refinements
 
   # Prompt can be short (apc>) or long (username@apc>)
@@ -10,25 +10,27 @@ class ApcOS < Oxidized::Model
     cfg.gsub("\r", "")
   end
 
-  cmd 'about' do |cfg|
+  cmd 'about', input: :ssh do |cfg|
     cfg = clean(cfg)
     cfg = cfg.reject_lines [/^Management Uptime: /, /^Date: /, /^Time: /]
     comment cfg
   end
 
-  cmd 'upsabout' do |cfg|
+  cmd 'upsabout', input: :ssh do |cfg|
     cfg = clean(cfg)
     comment cfg
   end
 
-  cmd 'detstatus -ss' do |cfg|
+  cmd 'detstatus -ss', input: :ssh do |cfg|
     cfg = clean(cfg)
     comment cfg
   end
 
-  cmd 'config.ini', input: :scp do |cfg|
+  cmd 'config.ini', input: %i[scp ftp] do |cfg|
     "; ========== config.ini ==========\n" + cfg
   end
+
+  inputs [:ssh, %i[scp ftp]]
 
   cfg :ssh do
     pre_logout 'exit'