+++ spec/model/apcaos_spec.rb
@@ -10,27 +10,34 @@ describe 'Model apc_aos' do
 
   it "fetches the configuration with ssh and scp" do
     @node = Oxidized::Node.new(name:     'example.com',
-                               input:    'ssh',
+                               input:    'ssh,scp',
                                output:   'file',
-                               model:    'apcos',
+                               model:    'apcaos',
                                username: 'alma',
-                               password: 'armud')
+                               password: 'armud',
+                               timeout:  1)
     Oxidized::SSH.any_instance.stubs(:connect).returns(true)
     Oxidized::SSH.any_instance.stubs(:node).returns(@node)
     Oxidized::SSH.any_instance.stubs(:connect_cli).returns(true)
     Oxidized::SSH.any_instance.stubs(:disconnect).returns(true)
     Oxidized::SSH.any_instance.stubs(:disconnect_cli).returns(true)
+    Oxidized::SCP.any_instance.stubs(:connect).returns(true)
+    Oxidized::SCP.any_instance.stubs(:node).returns(@node)
+    Oxidized::SCP.any_instance.stubs(:connect_cli).returns(true)
+    Oxidized::SCP.any_instance.stubs(:disconnect).returns(true)
+    Oxidized::SCP.any_instance.stubs(:disconnect_cli).returns(true)
+
     model = YAML.load_file('spec/model/data/apcos#SMT750IC_v2.5.0.8#custom_simulation.yaml')
 
     commands = ["about", "upsabout", "detstatus -ss"]
-    commands.each do |c|
-      c_result = model['commands']["#{c}\n"]
-      c_result = "\"#{c_result}\"".undump
+    commands.each do |cmd|
+      cmd_result = model['commands']["#{cmd}\n"]
+      cmd_result = "\"#{cmd_result}\"".undump
       Oxidized::SSH.any_instance.expects(:cmd)
-                   .with(c).returns(c_result)
+                   .with(cmd).returns(cmd_result)
     end
-    Oxidized::SSH.any_instance.expects(:cmd)
-                 .with("config.ini", input: :scp)
+    Oxidized::SCP.any_instance.expects(:cmd)
+                 .with("config.ini")
                  .returns(model['commands']['config.ini'])
 
     status, result = @node.run